replace index.vue
